requirements.txt
google-auth==2.16.0
pyperclip==1.8.2
